backend/requirements.txt
Django==5.2.10
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
django-crontab==0.7.1

celery==5.6.2
amqp==5.3.1
billiard==4.2.4
kombu==5.6.2
vine==5.1.0

redis==7.1.0
psycopg2-binary==2.9.11

requests==2.32.5
python-dotenv==1.2.1
python-dateutil==2.9.0.post0

asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
typing_extensions==4.15.0
six==1.17.0

certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3

click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt_toolkit==3.0.52
wcwidth==0.5.2
packaging==26.0
exceptiongroup==1.3.1
async-timeout==5.0.1

